pyproject.toml
[project]
name = "ai-agent2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.118.0",
    "inngest[ai]>=0.4.22",
    "llama-index-core>=0.14.3",
    "llama-index-readers-file>=0.5.4",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
